+++ api/src/main/java/net/warp_scores/warpscores/model/EntityType.java
@@ -2,5 +2,5 @@ package net.warp_scores.warpscores.model;
 
 
 public enum EntityType  {
-    League, Competition, Circuit
+    League, Competition
 }