+++ frontend/src/components/misc/Navigation.jsx
@@ -10,7 +10,7 @@ import prettyPrint from '../../util/prettyPrint';
 
 const { isProduction } = config;
 
-function Navigation({ currentPage, parentPage, currentLabel, league, competition, circuit, team, race, coach, circuitLeg, circuitLegEntity, leagueSystems, selectedLeagueSystemId, onSelectLeagueSystem }) {
+function Navigation({ currentPage, parentPage, currentLabel, league, competition, team, race, coach, leagueSystems, selectedLeagueSystemId, onSelectLeagueSystem }) {
   const intl = useIntl();
   const isPage = (pageName, currentPageName) => {
     return pageName === currentPageName;
@@ -18,16 +18,6 @@ function Navigation({ currentPage, parentPage, currentLabel, league, competition
   const leagueLink = league && league.length > 0 && league[0] ? `/league/${league[0]}` : '/';
   const competitionLink = competition && competition.length > 0 && competition[0] ? `/competition/${competition[0]}` : '';
   const teamLink = team ? `${competitionLink}/team/${team[0]}` : '';
-  //console.log("Circuit:",circuit," CircuitLeg:",circuitLeg,"CircuitLegEntity:",circuitLegEntity);
-  
-  const circuitLink = circuit ? `${isPage('admin', parentPage) ? '/admin' : ''}/circuit/${circuit[0]}` :
-  circuitLeg ? `${isPage('admin', parentPage) ? '/admin' : ''}/circuit/${circuitLeg[0].parts[0]}` : 
-  circuitLegEntity ? `${isPage('admin', parentPage) ? '/admin' : ''}/circuit/${circuitLegEntity[0].parts[0]}` : ''
-
-  ;
-  const circuitLegLink = circuitLeg ? `${circuitLink}/leg/${circuitLeg[0]}` : '';
-  const circuitLegEntityLink = circuitLegEntity ? `${circuitLegLink}/entity/${circuitLegEntity[0]}` : '';
-
   return (
     <Flex>
       <Breadcrumb fontFamily="bigStar" spacing={1}>
@@ -53,27 +43,6 @@ function Navigation({ currentPage, parentPage, currentLabel, league, competition
         {isPage('staffProfile', currentPage) && currentLabel && (
           <BreadcrumbItem isCurrentPage flexWrap>
             <BreadcrumbLink variant="menu">{currentLabel}</BreadcrumbLink>
-          </BreadcrumbItem>
-        )}
-        {circuit && (
-          <BreadcrumbItem isCurrentPage={isPage('circuits', currentPage)} flexWrap>
-            <BreadcrumbLink variant="menu" as={RouteLink} to={circuitLink}>
-              {circuit[1]}
-            </BreadcrumbLink>
-          </BreadcrumbItem>
-        )}
-        {circuitLeg && (
-          <BreadcrumbItem isCurrentPage={isPage('circuits', currentPage)} flexWrap>
-            <BreadcrumbLink variant="menu" as={RouteLink} to={circuitLegLink}>
-              {circuitLeg[1]}
-            </BreadcrumbLink>
-          </BreadcrumbItem>
-        )}
-        {circuitLegEntity && (
-          <BreadcrumbItem isCurrentPage={isPage('circuits', currentPage)} flexWrap>
-            <BreadcrumbLink variant="menu" as={RouteLink} to={circuitLink}>
-              {circuitLegEntity[1]}
-            </BreadcrumbLink>
           </BreadcrumbItem>
         )}
         {league && (